requirements.txt
Django>=4.2
strawberry-graphql[django]>=0.220.0
strawberry-graphql-django==0.62.0